+++ bridge/_files_integrations.PHP5.php
@@ -27,7 +27,7 @@ return [
     __DIR__ . '/../src/Integrations/Integrations/Laravel/LaravelIntegration.php',
     __DIR__ . '/../src/Integrations/Integrations/Lumen/LumenIntegration.php',
     __DIR__ . '/../src/Integrations/Integrations/Guzzle/GuzzleIntegration.php',
-    __DIR__ . '/../src/Integrations/Integrations/Psr/PsrIntegration.php',
+    __DIR__ . '/../src/Integrations/Integrations/Psr18/Psr18Integration.php',
     __DIR__ . '/../src/Integrations/Integrations/Yii/YiiIntegration.PHP5.php',
     __DIR__ . '/../src/Integrations/Integrations/Nette/NetteIntegration.php',
     __DIR__ . '/../src/Integrations/Integrations/WordPress/WordPressIntegration.php',
