+++ bridge/_files_integrations.php
@@ -30,7 +30,7 @@ return [
     __DIR__ . '/../src/Integrations/Integrations/Laravel/LaravelIntegration.php',
     __DIR__ . '/../src/Integrations/Integrations/Lumen/LumenIntegration.php',
     __DIR__ . '/../src/Integrations/Integrations/Guzzle/GuzzleIntegration.php',
-    __DIR__ . '/../src/Integrations/Integrations/Psr/PsrIntegration.php',
+    __DIR__ . '/../src/Integrations/Integrations/Psr18/Psr18Integration.php',
     __DIR__ . '/../src/Integrations/Integrations/Roadrunner/RoadrunnerIntegration.php',
     __DIR__ . '/../src/Integrations/Integrations/Yii/YiiIntegration.php',
     __DIR__ . '/../src/Integrations/Integrations/Nette/NetteIntegration.php',
