+++ ext/integrations/integrations.h
@@ -25,7 +25,7 @@
     INTEGRATION(PDO, "pdo")                     \
     INTEGRATION(PHPREDIS, "phpredis")           \
     INTEGRATION(PREDIS, "predis")               \
-    INTEGRATION(PSR, "psr")                     \
+    INTEGRATION(PSR18, "psr18")                 \
     INTEGRATION(ROADRUNNER, "roadrunner")       \
     INTEGRATION(SLIM, "slim")                   \
     INTEGRATION(SYMFONY, "symfony")             \
