+++ src/Integrations/Integrations/IntegrationsLoader.php
@@ -17,7 +17,7 @@ use DDTrace\Integrations\Nette\NetteIntegration;
 use DDTrace\Integrations\Pcntl\PcntlIntegration;
 use DDTrace\Integrations\PDO\PDOIntegration;
 use DDTrace\Integrations\Predis\PredisIntegration;
-use DDTrace\Integrations\Psr\PsrIntegration;
+use DDTrace\Integrations\Psr18\Psr18Integration;
 use DDTrace\Integrations\Slim\SlimIntegration;
 use DDTrace\Integrations\Symfony\SymfonyIntegration;
 use DDTrace\Integrations\Web\WebIntegration;
@@ -76,7 +76,7 @@ class IntegrationsLoader
         $this->integrations[MysqliIntegration::NAME] =
             '\DDTrace\Integrations\Mysqli\MysqliIntegration';
         $this->integrations[PsrIntegration::NAME] =
-            '\DDTrace\Integrations\Psr\PsrIntegration';
+            '\DDTrace\Integrations\Psr18\Psr18Integration';
 
         // Add integrations as they support PHP 8
         if (\PHP_MAJOR_VERSION >= 8) {
